Lower to 1x threads
