+++ geode-benchmarks/src/main/java/org/apache/geode/benchmark/tests/AbstractPartitionedFunctionBenchmark.java
@@ -15,40 +15,14 @@
 
 package org.apache.geode.benchmark.tests;
 
-import static org.apache.geode.benchmark.topology.ClientServerTopology.Roles.CLIENT;
 import static org.apache.geode.benchmark.topology.ClientServerTopology.Roles.SERVER;
 
-import benchmark.geode.data.BenchmarkFunction;
-
-import org.apache.geode.benchmark.LongRange;
-import org.apache.geode.benchmark.tasks.CreateClientProxyRegion;
 import org.apache.geode.benchmark.tasks.CreatePartitionedRegion;
-import org.apache.geode.benchmark.tasks.PrePopulateRegion;
-import org.apache.geode.benchmark.tasks.RegisterFunction;
-import org.apache.geode.benchmark.topology.ClientServerTopology;
-import org.apache.geode.perftest.PerformanceTest;
 import org.apache.geode.perftest.TestConfig;
 
-abstract class AbstractPartitionedFunctionBenchmark implements PerformanceTest {
-  private LongRange keyRange = new LongRange(0, 1000000);
-
-  public final void setKeyRange(LongRange keyRange) {
-    this.keyRange = keyRange;
-  }
-
-  public final LongRange getKeyRange() {
-    return keyRange;
-  }
-
+abstract class AbstractPartitionedFunctionBenchmark extends AbstractFunctionBenchmark {
   @Override
-  public TestConfig configure() {
-    TestConfig config = GeodeBenchmark.createConfig();
-    config.threads(Runtime.getRuntime().availableProcessors() * 4);
-    ClientServerTopology.configure(config);
+  protected void configureRegion(TestConfig config) {
     config.before(new CreatePartitionedRegion(), SERVER);
-    config.before(new CreateClientProxyRegion(), CLIENT);
-    config.before(new PrePopulateRegion(keyRange), SERVER);
-    config.before(new RegisterFunction(new BenchmarkFunction(keyRange)), SERVER);
-    return config;
   }
 }
